+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Improved suggestion for pipeline operator in malformed query (tolerant mode)
 - Added suggestion for object keys in array-like notation property accessor
 - Renamed node type `Property` to `ObjectEntry`
+- Added `ascN` and `descN` to define natural sorting functions
 
 ## 1.0.0-beta.2 (May 17, 2020)